wareIDs may be the empty string.
Zero wareIDs should be serialized as the empty string rather than a colon.

WareIDs with a type specified are clearly not empty, and so in this case if the hash is left empty we will standardize it to a string of dash.  This is not "necessary" but in case of odd usage or bugs tends to result in more sensible error messages in general.

Signed-off-by: Eric Myhre <hash@exultant.us>