As per in rio: answer our todo about serial errors.
I'm not *pleased* with this answer per se, but it works.

The choices that are reachable given the type system are roughly:

  1. do something very fancy with refmt on those interface fields
    - PRO: totally a valid option, realistically.
    - PRO: can do enum range check right at unmarshal time
    - CON: sucks if that's realistically a requirement for getting errcat adopted by anyone else, so we'd rather dogfood an approach with less dependencies.
  2. copypasta ErrStruct types and sub in specific category types
    - PRO: does the right things with stdlib unmarshal
    - MEH: doesn't do enum check
    - CON: copypasta!
      - can try to reduce with embeddable thing?  but still kind of losing game because the whole point is putting more type info on that category field.
    - CON: cannot easily assign to it!!  because all the error factory methods return `error` type!
      - and mind you this is not just a casting thing either.  you'd have to make a method for assigning that field, essentially.
  3. something about putting a correct struct type into the api type's `error` field
    - PRO: does the right things with stdlib unmarshal
    - MEH: doesn't do enum check
    - MEH: (as opposed to option 2, copypasta) this doesn't require copypasta, which is good (but like in a baseline way)
    - CON: this never results in graceful code: you'll almost certainly run into The Stupid thing in stdlib json where you have to dither the *parent* object with a custom behavior for its field.
      - it also doesn't generalize between serialization libraries very well, e.g. refmt makes no attempt to honor `JSONMarshal` funcs.
      - but this customized behavior is still easily hidden in an api package, in our figuring at least, so sort of okay i guess

What we've done here is Option 2, and for the moment we've opted for locating that "method for assigning that field" is in rio's main method.  This may be something to reconsider in the future, but for now we seem to have the interesting state of affairs that such objects will only be generated for the first time from... right there, and then deserialized again ever after; so, just for fun, let's see how long we can get away with keeping it out of wider sight.

AMEND: ported that main method stuff into here after all.  I want to be able to test such glue code from here, and also see the whole impact in this one commit diff.  This added the errcat lib to our list of deps for this repo, sadly, but that's a pretty minor drawback afaict.

This area remains open for improvements, as far as I'm concerned.  Notice how we're also still missing any enum range validity checks on the deserializing side, for example.

Signed-off-by: Eric Myhre <hash@exultant.us>