Try two diff libraries.  Neither acceptable.
The google struct one is probably neat... for structs.  It does not do breakdowns of strings, as far as I can tell, which makes it not so useful for large string compares as these API fixtures are.

The diffmatchpatch package is pretty neat in its interfaces, and I like it, but it doesn't do line-wise operation, which is sort of a "you had one job" thing in my perspective.  Perhaps I could hack enough layers around it to make that UI experience in the terminal nice enough to be useful, but... Bleh.  I want an off-the-shelf solution.

Another one which looks like a keeper coming up next commit.

Signed-off-by: Eric Myhre <hash@exultant.us>