Make a useful diff test.
Go 1.9 is now required, because we're going to start using the new `t.Helper` feature as we grow our new minimalized test utils here.

Also, immediately fix an omission bug in our atlas setup.  Lul!

Test is ported from the tests in the earlier hitch copy of the API packages... which we have not replaced with this submodule yet, but will, as soon as we get those tests ported over here now!

Signed-off-by: Eric Myhre <hash@exultant.us>