add filters to formula input section.
This is a somewhat belated addition; we've always known this isn't going anywhere.  (No idea why I was too lazy to put it in correctly earlier, tbh.  Probably because we hadn't resolved the serialization questions around filters when this section was first redrafted; now we have.)

Requires adding named type for the struct of each input value (so we can refer to the type to define the atlas entry for it).

Renamed the 'FormulaInputs' type to 'UnpackTree' based on realization while hacking on rio that the specification for how to unpack a tree of stuff on the host's disk plainly and how to set up a container's chroot is basically the same (surprise, surprise) -- we'll probably be using this UnpackTree type as function arguments in some rio internals.

Signed-off-by: Eric Myhre <hash@exultant.us>