add ware to formula serialization fixture.
Required backfilling the atlas setup for FilesetFilter.

Which in turn required backfilling boolean encode support to refmt XD, which is then also yanked up to date here.

And after all that, looking at the fixture as it appears now is enough to make me consider writing a custom transform for the FilesetFilter type, because ew.  Increasing indentation signals complexity, and that is not something I'm sure I want to be seeing.  (Yes, omitempty is not correctly supported in refmt yet either; but even if it was, I still visually dislike seeing increasingly deeply indented things here.)

Signed-off-by: Eric Myhre <hash@exultant.us>