add PlacementMode param to rio.UnpackFunc.
This is a new requirement discovered as we proceed in developing good caching behaviors in rio.  There's a fairly sizable "rationale" doc in that repo about caching design which may be of interest to the reader.

An earlier draft of this had *two* params: one for placement mode, and one bool for skipCache.  We've combined those -- "direct" now summarizes the only aggregate combination those two params which actually makes sense (for example, "mount" and skipcache simply do not make sense).

Signed-off-by: Eric Myhre <hash@exultant.us>