Add URLs for fetch and save to Formula.
With it, another atlas specification which ignores them again -- the Formula.SetupHash func uses this.

This is... very contentious.  One of our big lessons-learned from the first generation of repeatr design is that it was very difficult to explain to people that the silo URLs would be *removed* again for hashing and memoization.  Or perhaps it wasn't hard to explain; but it certainly wasn't obvious either.  And once people Got that, a lightbulb would go on over their heads, and suddenly the point of the system starts to make sense.

So, we've gone through a lot of work to *separate* URLs from the formula.  And here we are adding them back again.  "What gives?!", you might rightly ask.

We need one document which -- alone -- is sufficient to launch execution.  That, in effect, means "includes URLs".

We've whiteboarded and prototyped a lot of designs which try to crowbar the URLs even further away from sight, but none of them fly well.  (Making repeatr *require* hitch for running: problematic.  Exposing other naming layers in formulas so we have something other than input paths to join other URL-containing docs against: *incredibly* problematic ("naming"->merklecollapse).  Importing more global configs in host: possible, but it quickly becomes unclear where to draw the line, and also chips away at the "pastebin'able executable" feature.  Requiring multiple files: okay, how are they bound?  Filename relationships?  What happens when I want to `repeatr run <(clevertrick)`?  Wrecked?  How much trouble are we in if we need to write the two split files atomically and not have race issues?  And so on: no, that way lies madness.)

Thus, here we are: URLs are back in.  On a positive note, though, we've at least got them more clearly physically separated in the document: the inputs section is substantially more compact; and more importantly, when explaining what parts are hashed for memoization and identifying identical build setups, we can simply hold up a hand covering the entire bottom of the document.

The URL sections may be quite redundant.  (If you're using Reppl right now with the previous generation of Repeatr, you may have noticed the incredible repetition of "file+ca://./wares/" that shows up after a while.)  We're going to try simply biting the bullet on that.  The discovery of the problem with the Reppl prototype also discovers its own solution: *formulas are simply becoming more like assembly*; in the big picture, users rarely write them directly anymore.  So, our "planner" family of tools can invent arbitrarily fancy ways to make URLs non-irritating (and use hitch to do large amounts of the work, too), and they'll simply serialize that flattened plan of warehouse URLs into the formula, and users don't need to see this unless they're diving deep into debugging.  Hopefully this works out well; if not, we'll reconsider accordingly.

Signed-off-by: Eric Myhre <hash@exultant.us>