Add Formula.Clone helper method.
We'll strongly desire this momentarily for adding a bunch more tests.  (It's funny; goconvey's tree-forking methodology saved us from a loooot of needs of this sort of thing, in other projects where we'ved used that.)

The clone method shells out to refmt.  (Which we yank in an update for, in order to get these handy methods.)

A quick test is included against reflect.DeepEqual as a sanity checker.  (It'll call BS if we ever forget to add fields to the atlas entries, etc.)

Signed-off-by: Eric Myhre <hash@exultant.us>