repeatr: yank func APIs up to date.
- Bring in all the error serialization helpers prototyped in Rio;
- Declare fresh types for logging;
- Fix the presense of warehouses in func args (this changed when we merged that last branch about urls now being in formulae after all);
- Use a monitor object indirection like Rio does (makes it much easier to have DTRT zero values);

... yeah, that about sums it up.

We have a missing concept for how to hand in inputs (it could be as simple as an io.Reader, but I'm not sure if we want to do that, or a channel which makes chunking control more explicit): todo.

We'll need helpers for making output DTRT, because nobody really wants to deal with that event stream.  But I think that's okay: we end up needing such output buffering control proxies *anyway*, in order to deal with things like log events and user output with "\r" content not coliding to make a mess.

It's not clear if this func will be able to do doubletime for repeatr's internal code concept of executors as well as the wrapper for the entire tool.  Maybe.  (But more likely is that our previous abstraction of executors was just kinda not very correctly separated... but hey, the utility of a 'demo' executor -- which does, yes, need to see the whole inputs and outputs specs -- makes things muddy.)

Signed-off-by: Eric Myhre <hash@exultant.us>