repeatr: add InputControl.
Aside: considered using `[]byte` for input and output.  Refraining: stdlib json will not handle the serialization of such strings the way we mean.  It's better for the majority of typical usage to plug our ears and sing lalala to the truth of bytes, escaping them only when interesting ones appear.

Signed-off-by: Eric Myhre <hash@exultant.us>