repeatr: correction to saveUrls: only one entry.
We've learned from practical usage of the previous generation of repeatr that multiple output upload locations *as part of the formula run itself* are a really junk idea: what you want is in fact almost always to save it to one place, as quickly as possible, and (if a whole pipeline completes satisfactorily..!) later mirror up the whole set of results to as many servers as you want.  Critically, the mirroring operation is trivial to retry, whereas a tethering the retry of uploads to a retry of a whole build is kind of a strange way to waste time.

Signed-off-by: Eric Myhre <hash@exultant.us>