repeatr: separate contextual info from formula.
This is the (more or less promised) revisement to a576eb49248d09e834e967cc4d62a7bd6d9c7f53 ("Add URLs for fetch and save to Formula").

Formulas are now again a structure of purely the memoization-hash appropriate stuff.  The URLs are extracted to a separate top-level object.

And the serial format we expect from formula files changes: most of the content is now under the top-level object names "formula" and "context".

The `repeatr.RunFunc` signiture now takes the formula and its context as separate arguments (which makes sense, semantically); at the same time, we do still typically expect them to be serialized in the same actual file (for all the same still-valid reasoning outlined in commit a576eb4 regarding atomiticity and such practicalities).

The `api.FormulaCasAtlas` object could have been tossed back out, because it no longer needs to contain a *separate* atlas entry for the subset of `Formula` fields.  However, I'm going to keep it, because I kind of like having its minimialized whitelist of atlas entries there as a defense-in-depth against accidentally serializing anything bullshitty when computing the SetupHash.

Inspired-by: Calvin Behling <calvin.behling@gmail.com>
Signed-off-by: Eric Myhre <hash@exultant.us>