error table updates and source reorg.
The gluey bits for serialization now have their own files in each package.  You'll notice they're literally copypasta aside from the package header.  This is intentional.  (Golang!)

An error table for the rio package is added, and a function to get the rightful exit codes from a categorized error.  It is redundant with the table in the repeatr package.  I accept this.

I left the ErrorCategory defs and all their const declarations in the main file of API types in each package.  I want those files to be essentially the full documentation of the API surface.  (The exit codes are... ancilliary.  I'd leave them in the same file if there was a way to make them less obscenely verbose.)

Signed-off-by: Eric Myhre <hash@exultant.us>