add formula.Apply(formula) method.
This will be used in `repeatr twerk` for stacking on parameters
that are useful for interactive modes, so you can run things based
on a formula but with the necessary variations easily.

It'll also likely be useful in other higher level tools momentarily,
for example the skylark-based formula composition tools we're
experimenting with building out right now.

Signed-off-by: Eric Myhre <hash@exultant.us>