Introduce concept of "basting" groups of formulas.
This is named for Quilting: https://en.wikipedia.org/wiki/Quilting#Basting_the_layers
(The name will be subject to change, because it's a bit of a mixed metaphor
in comparison to all the rest of our nomenclature to date, but we'll roll
with it for now.)

The idea is that, like a quilt into which you put some loose, temporary
stitching on to hold things together while you work, a "basting" in the
Timeless APIs will hold together a bunch of formulas which rely on each
other to produce intermediate wares.

Basting will now be a reified concept, and not just something for hitch.
We'll actually teach Repeatr (or possibly another intermediate tool?) to
consume Basting, and compute the whole graph of builds at once.
This will likely simply *replace* the "replay" concept currently in the
hitch.go files, as it's almost isomorphic, we're just expanding the
definitions now.

Basting will probably be what's generated by almost all "layer 3" (see the
latest docs in the book from the Timeless doc repo) toolings in the future.
This makes its reponsbilities... somewhat complicated.  We'll likely see
Basting objects exist in various levels of completeness throughout those
layers of composition.

Signed-off-by: Eric Myhre <hash@exultant.us>