Give FormulaContext info a home in Basting.
As usual, we're just *frustrated* whenever we have to admit this impure,
not-timeless details into things.  But they've got to go somewhere.  And the
main lesson of our previous generation of API design is that they must *stay*
separate at all layers, or it will manufacture work for the next layer up.

Signed-off-by: Eric Myhre <hash@exultant.us>