Add Imports to FormulaUnion.
This is kind of controvertial.  But then, to be fair, so was making this union
at all and putting Context here in the first place.

The practical reason to do this is that as we're experimenting with composing
hunks of partial formula with Skylark in the Heft project, we want to compose
importable releaseID info as well.  And these simply go together all the time.

The controvertial bit is that Repeatr is using this struct too, and by design
it *does not care* about import names, and furthermore *will not check them*,
because by design it should not even know how to inquire about a hitch DB at
execution time... and therefore, even seeing the import information is arguably
kind of an overreach.

The choice here is going with basic practicality.  The overarching point of
this package is *not* to do The Cleverest Possible Things with the Golang type
system.  The aim is rather to be the simplest possible self-documenting code
which describes the overall shape of the structures.

And by that logic, Imports, Formula, and Context are often siblings.
Therefore, we'll represent them as such.  Even if that matrix is often
sparse.

Signed-off-by: Eric Myhre <hash@exultant.us>