Make FormulaUnion.Apply helper function.
This is really borderline on belonging in Heft.  I'm not sure it's that likely
to be reused in any other context.

Formula.Apply could arguably have originally been kept only in Repeatr very near
the CLI parts that might use it for stacking modifiers, by that logic, though.
And *that's* certainly already turning out to be reusable and useful.

Signed-off-by: Eric Myhre <hash@exultant.us>