hitch: traversal helper funcs.
This includes some error categories... but not an ErrorTable like
several other packages have already sprouted.  That's because these
helper functions are just in-memory utilities; not necessarily
intended to show up in an (potentially exec level) API boundary.
(Yet.  And it's a bit in flux where we actually expect hitch procs to
begin and end; the heft/commission research project seems to indicate
there's going to be more bleeding at the edges than earlier guessed.)

Signed-off-by: Eric Myhre <hash@exultant.us>