go-timeless-api, clean branch reboot.
Same readme, license, etc as before.

I want to do a cleanroom experiment again where we *start* with a
recursive-capable pipeline definiton, and see where we get from
there.  There will also be (many) more validators, possibly at the
expense of adding more types (for example, whether or not the hash
pin resolution is done is orthagonal to whether *every other field*
in that structure is present and how it's handled, so making a type
just to handle that distinction is just makework of no utility).

Signed-off-by: Eric Myhre <hash@exultant.us>