First pass sketch of recursive module pipeline.
We're going off a draft on paper so this is a tad hard to explain.

Rather immediately running into some amusing representational issues.
Like, almost all wiring tuples are two-tuples where part one is often
a zero value... except the top-level bindings of resolved names and
hashes, because those traded away their recursivity.  This is probably
correct; it's just funny-looking at first glance.  And terribly
upsettingly, since it's recursivity we're describing now, it's not a
tuple at all for such a reference; it's a slice.  Which is not
comparable for use as a map key.  So we'll have to have a string with
maintained duality semantics, hooray.

Signed-off-by: Eric Myhre <hash@exultant.us>