Attach util funcs to api types.
I've been trying to avoid "too much" code in the main API package,
because I'd like it to all be the serializable parts that are going to
remain universal even across languages.  That gets fincky, though.
Trying to then confine any of the other important shared validator
logic to a sub-package then requires things like these basic
contextualization helpers to be available there, and... meh.

Things got a lot less clear-cut when it became clear that proper
validators for some of these structures include nothing short of
graph walks.

Someday, we're going to make all this more concrete by having a literal
IDL schema of some kind, I hope.  Until then, I think I'm refining
my heuristic for minimal API package to be "only structs that are
serializable; and methods for validating and the most basic munging
may be attached".

Signed-off-by: Eric Myhre <hash@exultant.us>