cross-cutting motion towards pinning funcs.
- add basic hitch viewer interface back in (essentially unchanged)
- some more submodule reference helper funcs
- split the structure type for the top level module (i think embeds will do almost exactly everything we want here)
- a resolvePins tool which is driving us along (though yet unfinished)

aaaand it's time to review what we really meant with the whole imports vs importsResolved thing.
That dichotomic concept exists *somewhere* but it's tricky to decide exactly where it begins and should end.
See the comments in the diff -- tl;dr ingests are driving some of these design choices, but they... should
not be a crack in the dam that's allowed to widen without reason.

Signed-off-by: Eric Myhre <hash@exultant.us>