ResolvePins can now apply a ViewCatalogTool.
An ingestTool is still an entirely TODO concept, and will be needed soonish,
but perhaps we can get to a nice testable resting point before that feature.

I'm a liiiittle perturbed about the SubmoduleSlotReference type upon further
reflection.  If all the fields are purely dot-separated, it'll be ambiguous
to parse without context of the surrounding structure (and even then,
ambiguous if a step and an imported slot had the same name, which is
both admittedly a strange case to defend and yet *shouldn't* be a problem).
And also, I can't actually at this moment see a use where the StepName
should actually be non-zero; we're always using this to refer to imported
slots so far.  This might require review later; right now it's not actively
a problem yet, so I guess I'm going to leave it until more info turns up.

Signed-off-by: Eric Myhre <hash@exultant.us>