Operation atlasing, and test fixtures.
This includes a refmt bump because... *fanfare* We're using struct
map keys in the Operation.Inputs map, which is only supported just
moments ago: https://github.com/polydawn/refmt/pull/29 .  (Kinda
excited about that.)

Fixed an "off-by-one" of sorts in ParseSlotReference func.  Add a
notch to the roster: yes, testing your serialization literals is
a high value thing to do.

Signed-off-by: Eric Myhre <hash@exultant.us>