Module serial fixture -- recursive and with union!
This is really quite exciting.  When I started writing this iteration
of the schema types, the features to support cleanly serializing and
deserializing these things into union-style interfaces *didn't exist*.
And now it does.  And it fits perfectly.
https://github.com/polydawn/refmt/pull/30 is going to be a really
dramatic quality-of-life improvement for basically all code I have
that has a union anywhere in it... which is yes, *most* APIs.

The JSON fixture is a tad heavily indented.  This is the truth of the
matter, though: in the deeply recursing parts, it's consistently
cycling between fieldName, mapKey, typeHint; and none of those can be
elided without making the behavior incorrect.  And it's also just the
truth of things that this is a deeply recursive fixture on purpose!
(It's probably *not* going to be typical for humans to write more than
a single layer of submodule in practice unless they're already using
a composition engine that's going to provide a terser view than this.)

Wee punt on "ingest" type references.  Easy to fill in later.  It's
just another string-futzer, and the features it's describing need
some development before it's useful to push on the API part
(especially because we're anticipating some sort of "plugin" system
in that area, in order to deal with things like e.g. git without
going nuts or having bananas dependencies in the core, and plugins
mean nontriviality).

Signed-off-by: Eric Myhre <hash@exultant.us>