Contextualize function names.
Hopefully the cardinality of operations that build up in this package stays low
enough that this is organizationally sufficient; I'd rather not have a ton of
more specific sub-packages, because in the big picture, the consumers of the
API as a lib already tend have sub-packages matching these feature areas, and
the naming confusion would be non-fun.

Besides, all the stuff so far was implicitly "Catalog*" anyway, and that was
clearly bunkum, and that's now fixed.

Signed-off-by: Eric Myhre <hash@exultant.us>