rn Module.Operations -> Steps.
Serial form was already changed to "steps"; reason for this was that while
writing serial fixtures, it became clear that "operations" stuttered a lot
and looked much more confusing to the eye when one of the main members of
the union inside that field was "operation".  Just carrying it through, now.

Signed-off-by: Eric Myhre <hash@exultant.us>