Add warehouse info types.
This is a tad placeholder and not yet exercised.

We're angling towards this WareSourcing type being something that's
easier to curry along, and be closer associated with the catalog
types rather than a formula types -- the thought being that when
warehouse info is making to repeatr, it really is quite transient;
and therefore keeping all the organization oriented on the more
long-term storage half of things is gonna make a lot more sense.

The intent is to pretty much replace the old formulaContext system.
We've gone through a couple drafts on that already, and it's...
yeah, this last one wasn't perfect either.  Good thing we knew it,
and that nothing about how we handle these warehouse URLs has ever
been a factor in *any* of our hashes.

Signed-off-by: Eric Myhre <hash@exultant.us>