repeatr API.
Both the new API with RunFunc signiture and new types, AND
the entire implementation which flips it to the v200 types
(e.g. paths as input and output keys, no slot names, and
oldstyle 'formulacontext' for the warehouse URLs).

The warehouse URLs are placeholder hacks for now.
We'll need a more real story for that soon, but this is enough
to build some more really neat demos with.

(This is all folding into some new tools that are actually
implementing real runs of the new module format... and it's
getting pretty exciting, so long as you ignore some speedbumps
like this URL issue that we're continuing to kick the can on
for a while longer yet.)

Signed-off-by: Eric Myhre <hash@exultant.us>