repeatr client: port printing utils to be shared.
These are all the prettyprinters currently in repeatr; those will be
removed and pointed to these in the future.

It seems useful to extract them to our shared libs, because in general
there needs to be *one* process in control of the terminal at any time
(in order to make sure output is sequenced correctly and not
doesn't get broken up and wrongly interleaved); and therefore everyone
else has to forward events to it API-wise; and therefore we ought to
make the code for human-prettyprinting events available to any project,
since which process is in control of the terminal varies depending on
what layer of the stack you're working with.

I'll probably do a second round of refactors to these to make them
each their own functions, which will make it easier to say, e.g.,
"give me job output, but not container setup logs" -- but later.

Signed-off-by: Eric Myhre <hash@exultant.us>