Expose 'wareStaging' param in repeatr.Run API.
And introduce a type for that.

And docs for both the new WareStaging type as well as backtracking to add docs
for the WareSourcing type, which is both similar and yet not at all the same.

This is beginning to come full-ish circle back to the "formula context" types
in the earlier repeatr APIs.  It's a bit different, though: we've made
something that's re-usable for multiple formulas, and that's important.

What's interesting is that "reusable for multiple formulas" and "allow the
user to use the 'friendly' footgun of non-CA output warehouses" are...
in almost complete conflict, as the docs now expound upon.  Perhaps we'll
have to retain both after all (for the repeatr layers, at least; I don't
see any way that the higher level pipeline tools will ever re-expose the
details of per-operation-output-path warehouses).

Signed-off-by: Eric Myhre <hash@exultant.us>