FilesetFilter port.
These are not yet visibly used here, but this is the first step in getting an
updated rio API client here in the redux branch; first to use in stellar
(because I want to start accessing pack features directly up here) and likely
also to propagate back down to a new build of rio with some improvements.

Taking the opportunity to make a few changes, some of them to the
implementation pattern interally and some to the actual semantics:

- *two* types.  pack filters and unpack filters should be distinct; they do not always have the same valid enums and should not be freely substituted.
  - things like "mine" which were only ever valid in unpack mode are now more clearly so.
- when talking about unpacking, we now say "follow" instead of "keep".
- mtime can be configured for "now" mode on unpack (though I don't know why you'd want this).
- sticky and setuid/setgid are distinct now (phew).  you can follow sticky bits while rejecting setid bits.
- reactions to device nodes can also be configured.  (previously they were always kept, and if unpack tried to handle them when it didn't have privs, you'd just get a crashy error.)
- setid and dev now support "reject" options, where if they're encountered, rather than either keep/follow or zeroing them, we can explicitly halt outright.
  - this is useful on pack, in particular, if you want to avoid creating priv-needed-to-unpack things by accident.  which is a sane default!

And the new serialized representation is now overall a string with some very
simple parsing rules: it looks roughly like the "options" string for
filesystems that you're used to from /etc/fstab.  And this is coincidence,
more or less, but also unsurprisingly convergent evolution: this string is
describing almost exactly the same order of magnitude of complexity (a finite
list of values, each of which has a sum type of some unique enums and sometimes
one large cardinality range, but never anything recursive, etc).

It's not often that we actually *use these features* at all -- most of repeatr
and stellar don't actually *expose* filters to the user at all, though rio CLI
flags do.  Stellar might sprout an ingest:pack feature soon which will be a new
usecase for pack filters, though.

Signed-off-by: Eric Myhre <hash@exultant.us>