Back out on using WareSourcing in Rio levels.
I implemented the entire refactor and then it just turned out to be...
really surprisingly janky.  If we actually had rio commands that did
whole trees of operations in one go, it might be one thing; but that's
not how any of the Rio APIs work right now, so providing warehouse
location information in something designed to be a batch/bulk form just
doesn't make very much sense.

It's possible that the rio CLI commands will learn about wareSourcing,
and use that style of configuration when being used directly by users.
But when Rio is being driven around as a subsystem by some other part
of the stack, things make a lot more sense if we just assume the upper
layers of the stack knew what they were doing and can just feed an
already linear list of warehouses to check.

Signed-off-by: Eric Myhre <hash@exultant.us>