Rio exec client.
Previously the code for this lived in the Rio repository itself.
We're moving it here because it's easier to use without importing that
whole repository, and there's more than one project wanting to do so.

The tests become somewhat... less, as a result of this :/  The previous
home of this code in the Rio repo meant it could easily use a bunch of
Rio test utilities to get a ton of good coverage quickly.  That's
harder to replicate now, unfortunately.

I don't think the error types are particularly well persisted by this
code either, but then, neither are the ones in the repeatr exec client
at present, so let's just come back to both of those later.  (I've been
on a refactor binge to try to get everything on the same dang API
branch for more days in a row than I'd like, now; I desperately just
want everything to get up to one consistent level so the rest of these
refactors can be more incremental and less harried again.)

Signed-off-by: Eric Myhre <hash@exultant.us>