Pivoting several names to Formula rather than Op.
It's time to work on getting the repeatr master branch onto this commit chain
in the API currently known as "redux".  As I work on this, I'm increasingly
realizing that a bunch of concepts have been captured by what works well for
stellar, and not what necessarily needs to be the range of things supported
by repeatr, etc.  So there's going to be some refactors here.  This will be
the first and most trivial: just some naming cleanup.

We're going to keep with the current "Formula" nomenclature.  Anything that
does exist in formulas gets a name indicating as much; and if those things are
reused in the stellar layer objects, that's fine, but they're keeping their
names based on the lowest level of the stack that they first appeared in.

The repeatr exec client package broke in this commit.  I'll get back to you.
Finishing migrating that makes the diff blow out in size immensely.

Signed-off-by: Eric Myhre <hash@exultant.us>