Return of FormulaContext; ready for Repeatr to branchjump.
FormulaContext and some other types are publicly declared again.
Comments about their deprecation and replacement schedule previously whispered
in the repeatr client package turned out to be incorrect, and are removed.

FormulaRunRecord and OperationRecord now both exist.  They're *extremely*
similar, but nonetheless, both exist; they have different purposes.

The Repeatr exec client package got a lot shorter, because more of its types
are now raised to the publicly exported packages and reused.

There was a lot of deliberation behind this commit.  I actually wrote this
whole migration twice: the first time, I actually tried to switch the entirety
of the repeatr.RunFunc interface over to WareSourcing, WareStaging, and even
OperationResult -- destroying the older (Formula)RunRecord and FormulaContext
concepts.  This _didn't work_.  It didn't just blow out in scope; it
increasingly revealed itself as A Bad Idea.

The Repeatr system indexes everything by path.  It does this for a reason:
because paths are materially important.  Slotnames and references from the
Stellar/pipelining/modules layer *are not* "materially important"; so this
part of the design is certainly sturdy.  Trying to then make the warehouse
URL systems around the Repeatr layer indexed by something *different* than the
entire rest of the Formula just made things more complicated, more confusing,
less attractive, and generally served no purpose whatsoever.

Furthermore, it's *a feature* that Repeatr can accept more explicit systems of
warehousing (for outputs, especially) than Stellar/pipelining will readily
allow the user to configure.  Just because the higher level tools discourge
certain kinds of usage for their own reasons... does not imply Repeatr should
be so constrained: it is designed with the intent to be usable as a component,
even in other systems with radically different design constraints.
Therefore, we won't be expecting FormulaContext to go away any time soon.

And in light of all that, trying to have RunFunc return an OperationRecord
(that is, something keyed by SlotName, rather than path) would be just
obviously nonsense, so we can elide further discussion of that.

So!  The RunFunc is now (returned to) what *Repeatr's* API wants, and is not
being captured by any of the concerns specific to Stellar/pipelining/modules.
That functionality is now moved to a RunOperation method, which will still do
all the same sorts of conversion from named-references Operation to Formula,
and then on the way back will convert FormulaRunRecord to OperationRecord.
This thus is definitely an API break for Stellar, but also, the fix is easy:
just switching it to use the new method that does the same stuff, but applies
all the translations as a decorator on top of the repeatr.RunFunc API.
And as a bonus, the "bind inputs from map[SlotRef]WareID" feature is now
exported somewhere reusable.

Signed-off-by: Eric Myhre <hash@exultant.us>