Reintroduce FormulaRunRecord option metadata, and FormulaID field.
It's possibly that some of these will change some day -- I'm not sure
if the "hostname" as a blessed field would survive review today, rather
than just being kicked into "put that in the metadata grabbag".
But that's a discussion for some other time.  We're getting repeatr
onto this branch; and changes like that are not the present topic.

Signed-off-by: Eric Myhre <hash@exultant.us>