Improve error reporting in exec clients.
Returning the EOF error in that branch was a typo; the code from the
cmd.Wait is much more informative.

Although I'm not sure it includes as much context from stderr as I
would've liked...?  I thought that was added a few stdlib versions ago.
Maybe maybe it's not on this particular API path somehow.  Dunno.

Signed-off-by: Eric Myhre <hash@exultant.us>