Fileset*Filter: clearly flag when uninitialized.
We put all sorts of work in to make sure the Parse methods, etc, would
only ever return validly initialized values.....

... but it's still WAY to easy to accidentally get a "zero" value of
the struct sneaking in somewhere, and that desperately needs to be
identifiable.

Turns out this happens even against your best efforts when serialized
messages have the struct in them.  The custom transform funcs don't get
called if there are no tokens in the stream that cause the unmarshaller
to look at that field!  Sigh.

Sometimes I would give an arm and a leg to be able to provide a custom
definition of the zero value for one of my types.

Signed-off-by: Eric Myhre <hash@exultant.us>