Flip Operation.Inputs to Path:SlotRef ordering.
Previously, Operation.Inputs was in SlotRef:Path ordering.
Putting the SlotRef first scored well for consistency in one regard:
it meant all appearances of a SlotRef were on the *left* in a Module.

However, this was *inconsistent* in another almost every other regard:

- SlotRef:Path ordering of input is in "Value=:Assignment" order.
Every other piece of specification -- import mappings, output mappings,
and export mappings -- in a Module goes in the other way around:
"Assignment:=Value" order.  (This right-to-left "Assignment:=Value"
order is also typical in the majority of modern programming languages.)
Flipping makes assignment order consistent throughout the Module spec
(even if at the cost of consistency in slotref-on-the-left order).

- Repeatr Formulas, one layer down in the API layer model, are
in "Path:Hash" order.  Having the Path on the *left* in Formula.Input
while on the *right* in Operation.Input was staggeringly inconsistent.

- Having the Path as the map *key* is simply *more correct*.
It is invalid to have more than one input mapping into the same path
(see https://github.com/polydawn/stellar/issues/2 !), so we might
as well use the uniqueness of map keys to make this invalid concept
inexpressible.

Most of the diff is simply test fixture updates.

This will, of course, be a breaking change for serialized modules.
We'll have to bite the bullet on that.  It's a pretty quick and
mechanical rewrite, fortunately.

Signed-off-by: Eric Myhre <hash@exultant.us>